cleanup orderbook CLI
